+++ packages/sitecore-jss-angular/src/components/text.directive.spec.ts
@@ -215,7 +215,9 @@ describe('<span *scText />', () => {
       fixture.detectChanges();
 
       const rendered = de.nativeElement.innerHTML;
-      expect(rendered).toContain('<span>[No text in field]</span>');
+      expect(rendered).toContain(
+        '<span sc-default-empty-text-field-editing-placeholder="">[No text in field]</span>'
+      );
     });
 
     it('should render custom empty field component when provided, when field value is empty', () => {
@@ -281,7 +283,9 @@ describe('<span *scText />', () => {
           comp.field = field;
           fixture.detectChanges();
 
-          const fieldValue = de.query(By.css('sc-default-empty-text-field-editing-placeholder'));
+          const fieldValue = de.query(
+            By.css('span[sc-default-empty-text-field-editing-placeholder]')
+          );
           const metadataOpenTag = fieldValue.nativeElement.previousElementSibling;
           const metadataCloseTag = fieldValue.nativeElement.nextElementSibling;